experiments/requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cvat-cli==2.20.0
cvat_sdk==2.20.0
cycler==0.12.1
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.13.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
tinydb==4.8.0
toml==0.10.2
torch==2.4.0
torchsummary==1.5.1
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
triton==3.0.0
tuspy==0.2.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.81
ultralytics-thop==2.0.5
urllib3==2.2.2
wandb==0.17.7
watchdog==4.0.2